Homework-2-Basic-Control_Structures
Besides using the if/else statement, I used the Switch statement, made array from the signs and used it to access the result with less code.
Also I made an object from the signs with the possible value as a key and the actual sign as a value.
Bonus: Made a function to show A or AN in front of the sign, depending if the sign as a word is a vowel or consonant and I called it in the console.